+++ src/api/useAuth.jsx
@@ -1,8 +1,11 @@
 import { useEffect, useState } from 'react';
-import { auth } from './config.js';
-import { GoogleAuthProvider, signInWithPopup } from 'firebase/auth';
-import { addUserToDatabase } from './firebase.js';
 import { useNavigate } from 'react-router-dom';
+import {
+	addUserToDatabase,
+	GoogleAuthProvider,
+	signInWithPopup,
+} from 'firebase';
+import { auth } from './config.js';
 
 /**
  * A button that signs the user in using Google OAuth. When clicked,
