+++ src/components/index.js
@@ -1,6 +1,6 @@
+export * from './AddItem';
+export * from './Dialog';
+export * from './Emoji';
 export * from './ListItem';
 export * from './SingleList';
-export * from './Dialog';
-export * from './AddItem';
 export * from './SortedItemsMap';
-export * from './Emoji';
